refactor(wow): restructure modules/wow/ into folder module with mixins
Split the monolithic wow.py (1762 lines) into character/news/crafting
mixins composed via a single GroupCog, move crafting_order.py views into
modules/wow/views/, and relocate utils/blizzard.py to modules/wow/api.py.
