+++ NerdyPy/models/wow/__init__.py
@@ -7,6 +7,10 @@ from models.wow.crafting import (
     BIND_ON_EQUIP,
     BIND_TO_ACCOUNT,
     CURRENT_BOARD_VERSION,
+    ORDER_STATUS_CANCELLED,
+    ORDER_STATUS_COMPLETED,
+    ORDER_STATUS_IN_PROGRESS,
+    ORDER_STATUS_OPEN,
     RECIPE_TYPE_CRAFTED,
     RECIPE_TYPE_HOUSING,
     CraftingBoardConfig,
@@ -25,6 +29,10 @@ __all__ = [
     "BIND_ON_EQUIP",
     "BIND_TO_ACCOUNT",
     "CURRENT_BOARD_VERSION",
+    "ORDER_STATUS_CANCELLED",
+    "ORDER_STATUS_COMPLETED",
+    "ORDER_STATUS_IN_PROGRESS",
+    "ORDER_STATUS_OPEN",
     "RECIPE_TYPE_CRAFTED",
     "RECIPE_TYPE_HOUSING",
     "CraftingBoardConfig",
