+++ NerdyPy/modules/wow/news.py
@@ -1,94 +1,48 @@
 # -*- coding: utf-8 -*-
+"""WoW guild news cog: activity feed, mount tracking, background polling."""
 
 import asyncio
 import itertools
 import json
 from datetime import UTC, datetime, timedelta
-from enum import Enum
-from typing import LiteralString
 
-import discord
-from blizzapi import Language, Region, RetailClient
-from discord import Color, Embed, HTTPException, Interaction, TextChannel, app_commands
+from discord import Embed, HTTPException, Interaction, TextChannel, app_commands
 from discord.app_commands import checks
 from discord.ext import tasks
-from discord.ext.commands import Cog, GroupCog
-from models.wow import (
-    CURRENT_BOARD_VERSION,
-    CraftingBoardConfig,
-    CraftingOrder,
-    CraftingRoleMapping,
-    WowCharacterMounts,
-    WowGuildNewsConfig,
-)
-from utils.blizzard import (
-    CRAFTING_PROFESSIONS,
+
+from models.wow import WowCharacterMounts, WowGuildNewsConfig
+from modules.wow.api import (
     RateLimited,
     build_account_groups,
     check_rate_limit,
     clear_character_failure,
     get_asset_url,
-    get_best_mythic_keys,
-    get_profile_link,
-    get_raiderio_score,
     make_pair_key,
     parse_known_mounts,
     record_character_failure,
     should_skip_character,
     should_update_mount_set,
 )
-from utils.cog import NerpyBotCog
-from utils.errors import (
-    NerpyInfraException,
-    NerpyNotFoundError,
-    NerpyPermissionError,
-    NerpyUserException,
-    NerpyValidationError,
-)
+from modules.wow.characters import COLOR_ACHIEVEMENT, COLOR_ENCOUNTER, COLOR_MOUNT
+from utils.errors import NerpyNotFoundError, NerpyPermissionError, NerpyUserException, NerpyValidationError
 from utils.helpers import notify_error, register_before_loop, send_paginated
 from utils.permissions import validate_channel_permissions
 from utils.strings import get_string
-
-
-class WowApiLanguage(Enum):
-    """Language Enum for WoW API"""
-
-    DE = "de_DE"
-    EN = "en_US"
-    EN_GB = "en_GB"
-
-
-# Embed colors for guild news notifications
-COLOR_ACHIEVEMENT = Color.gold()
-COLOR_ENCOUNTER = Color.red()
-COLOR_MOUNT = Color.purple()
-COLOR_ITEM_LINK = Color(value=0x0099FF)  # WoW blue item link color
 
 # Stale character cleanup: remove mount data for characters gone from roster after this many days
 STALE_DAYS = 30
 
 
-@app_commands.checks.bot_has_permissions(send_messages=True, embed_links=True)
-@app_commands.guild_only()
-class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
-    """World of Warcraft API"""
+class WowNewsMixin:
+    """Guild news commands and background polling.
+
+    Mixed into the WorldofWarcraft GroupCog via __init__.py.
+    """
 
     guildnews = app_commands.Group(name="guildnews", description="manage WoW guild news tracking", guild_only=True)
-    craftingorder = app_commands.Group(name="craftingorder", description="manage crafting order board", guild_only=True)
 
-    def __init__(self, bot):
-        super().__init__(bot)
-        self.config = bot.config
-        self.client_id = self.config["wow"]["wow_id"]
-        self.client_secret = self.config["wow"]["wow_secret"]
-        self.regions = ["eu", "us"]
-
-        # Realm cache: "slug-region" -> {"name": "Blackrock", "region": "eu", "slug": "blackrock"}
-        self._realm_cache: dict[str, dict] = {}
-        self._realm_cache_lock = asyncio.Lock()
-
-        # Guild news config (optional section with defaults)
-        gn_config = self.config["wow"].get("guild_news", {})
+    def _init_news(self, bot):
+        gn_config = bot.config["wow"].get("guild_news", {})
         self._poll_interval = gn_config.get("poll_interval_minutes", 15)
         self._mount_batch_size = gn_config.get("mount_batch_size", 20)
         self._track_mounts = gn_config.get("track_mounts", True)
@@ -97,330 +51,6 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
         register_before_loop(bot, self._guild_news_loop, "Guild News")
         self._guild_news_loop.change_interval(minutes=self._poll_interval)
         self._guild_news_loop.start()
-
-        register_before_loop(bot, self._crafting_cleanup_loop, "Crafting Cleanup")
-        self._crafting_cleanup_loop.start()
-
-    async def cog_load(self):
-        # Ensure tables exist on existing databases before running migration checks.
-        self.bot.create_all()
-        self.bot.loop.create_task(self._run_board_migrations())
-
-    def cog_unload(self):
-        self._guild_news_loop.cancel()
-        self._crafting_cleanup_loop.cancel()
-
-    async def _run_board_migrations(self):
-        """Wait for bot ready then migrate any stale crafting boards."""
-        await self.bot.wait_until_ready()
-        await self._migrate_boards()
-
-    async def _migrate_boards(self):
-        """Upgrade crafting boards that are behind CURRENT_BOARD_VERSION."""
-        with self.bot.session_scope() as session:
-            from sqlalchemy import or_
-
-            stale = (
-                session.query(CraftingBoardConfig)
-                .filter(
-                    or_(
-                        CraftingBoardConfig.BoardVersion.is_(None),
-                        CraftingBoardConfig.BoardVersion < CURRENT_BOARD_VERSION,
-                    )
-                )
-                .all()
-            )
-            boards = [(c.Id, c.GuildId, c.ChannelId, c.BoardMessageId, c.BoardVersion) for c in stale]
-
-        await asyncio.gather(
-            *[
-                self._migrate_single_board(config_id, guild_id, channel_id, message_id, version)
-                for config_id, guild_id, channel_id, message_id, version in boards
-            ]
-        )
-
-    async def _migrate_single_board(
-        self, config_id: int, guild_id: int, channel_id: int, message_id: int | None, version: int
-    ):
-        """Migrate a single crafting board to CURRENT_BOARD_VERSION.
-
-        On success or terminal LookupError (guild/channel/message permanently gone),
-        bumps BoardVersion so the board isn't retried on every restart. Transient
-        discord.HTTPException (429, 5xx) does NOT bump the version so the migration
-        is retried on the next startup.
-        """
-        from modules.views.crafting_order import CraftingBoardView
-
-        lang = self._lang(guild_id)
-
-        try:
-            guild = self.bot.get_guild(guild_id)
-            if not guild:
-                raise LookupError(f"guild {guild_id} not in cache")
-
-            try:
-                channel = guild.get_channel(channel_id) or await guild.fetch_channel(channel_id)
-            except (discord.NotFound, discord.Forbidden) as exc:
-                raise LookupError(f"channel {channel_id} inaccessible") from exc
-
-            if not message_id:
-                raise LookupError("no board message id")
-
-            try:
-                msg = await channel.fetch_message(message_id)
-            except (discord.NotFound, discord.Forbidden) as exc:
-                raise LookupError(f"board message {message_id} inaccessible") from exc
-            if msg.author.id != self.bot.user.id:
-                raise LookupError("board message not authored by bot")
-
-            # v1 → v2: add housing button
-            embed = msg.embeds[0] if msg.embeds else discord.Embed(color=discord.Color.gold())
-            new_view = CraftingBoardView(
-                self.bot,
-                label=get_string(lang, "wow.craftingorder.create_button"),
-                housing_label=get_string(lang, "wow.craftingorder.housing_button"),
-            )
-            try:
-                await msg.edit(embed=embed, view=new_view)
-            except (discord.NotFound, discord.Forbidden) as exc:
-                raise LookupError(f"board message {message_id} edit failed") from exc
-            self.bot.log.info(
-                "Board migration v%d→v%d: guild=%d config=%d", version, CURRENT_BOARD_VERSION, guild_id, config_id
-            )
-
-        except LookupError as exc:
-            # Terminal failure (guild/channel/message gone permanently) — bump version so
-            # we don't retry on every restart. Notify the channel if it's still reachable.
-            self.bot.log.warning(
-                "Board migration failed (terminal) for config=%d guild=%d: %s", config_id, guild_id, exc
-            )
-            try:
-                guild = self.bot.get_guild(guild_id)
-                if guild:
-                    channel = guild.get_channel(channel_id) or await guild.fetch_channel(channel_id)
-                    await channel.send(
-                        get_string(
-                            lang,
-                            "wow.craftingorder.board_migration_failed",
-                            channel=channel.mention,
-                            guild=guild.name,
-                        )
-                    )
-            except discord.HTTPException as notify_exc:
-                self.bot.log.debug("Board migration: could not notify channel %d: %s", channel_id, notify_exc)
-
-        except discord.HTTPException as exc:
-            # Transient failure (429, 5xx) — log and skip version bump so migration retries next startup.
-            self.bot.log.warning(
-                "Board migration failed (transient) for config=%d guild=%d: %s — will retry", config_id, guild_id, exc
-            )
-            return
-
-        # Bump version on success or terminal LookupError.
-        with self.bot.session_scope() as session:
-            config = session.get(CraftingBoardConfig, config_id)
-            if config:
-                config.BoardVersion = CURRENT_BOARD_VERSION
-
-    async def _call_api(self, api_method, config_id, label, *args, rate_limited_event=None, stats=None, **kwargs):
-        """Call a Blizzard API method with standard rate-limit and error handling.
-
-        Returns the result on success, or None on failure (already logged).
-        Sets rate_limited_event and increments stats["skipped_error"] when provided.
-        """
-        self.bot.log.debug(f"Guild news #{config_id}: {label}")
-        try:
-            result = await asyncio.to_thread(api_method, *args, **kwargs)
-            check_rate_limit(result)
-            return result
-        except RateLimited:
-            self.bot.log.warning(f"Guild news #{config_id}: rate limited on {label}")
-            if rate_limited_event:
-                rate_limited_event.set()
-            return None
-        except Exception as ex:
-            log_fn = self.bot.log.debug if stats is not None else self.bot.log.warning
-            log_fn(f"Guild news #{config_id}: {label} failed: {ex}")
-            if stats is not None:
-                stats["skipped_error"] += 1
-            return None
-
-    # ── Realm cache & autocomplete ─────────────────────────────────────
-
-    async def _ensure_realm_cache(self):
-        """Lazily populate the realm cache from both EU and US regions."""
-        if self._realm_cache:
-            return
-
-        async with self._realm_cache_lock:
-            # Double-check after acquiring lock
-            if self._realm_cache:
-                return
-
-            cache = {}
-            for region in self.regions:
-                try:
-                    api = self._get_retailclient(region, "en")
-                    data = await asyncio.to_thread(api.realms_index)
-                    check_rate_limit(data)
-                    for realm in data.get("realms", []):
-                        slug = realm.get("slug", "")
-                        name = realm.get("name", slug)
-                        if slug:
-                            cache[f"{slug}-{region}"] = {"name": name, "region": region, "slug": slug}
-                except Exception as ex:
-                    self.bot.log.warning(f"Failed to fetch realm index for {region}: {ex}")
-
-            self._realm_cache = cache
-            self.bot.log.info(f"Realm cache populated with {len(cache)} entries")
-
-    # noinspection PyUnusedLocal
-    async def _realm_autocomplete(
-        self, interaction: discord.Interaction, current: str
-    ) -> list[discord.app_commands.Choice[str]]:
-        """Autocomplete callback for the realm parameter."""
-        await self._ensure_realm_cache()
-
-        current_lower = current.lower()
-        matches = []
-        for key, info in self._realm_cache.items():
-            if current_lower in info["name"].lower() or current_lower in info["slug"]:
-                label = f"{info['name']} ({info['region'].upper()})"
-                matches.append(discord.app_commands.Choice(name=label, value=key))
-            if len(matches) >= 25:
-                break
-
-        return matches
-
-    # ── Shared helpers ──────────────────────────────────────────────────
-
-    async def _parse_realm(self, realm: str, lang: str = "en") -> tuple[str, str]:
-        """Parse a realm string like 'blackrock-eu' into (realm_slug, region).
-
-        Validates against the realm cache if populated. Plain slugs default to EU.
-        """
-        realm = realm.lower()
-        if "-" in realm:
-            parts = realm.rsplit("-", 1)
-            if parts[1] in self.regions:
-                realm_slug, region = parts[0], parts[1]
-            else:
-                realm_slug, region = realm, "eu"
-        else:
-            realm_slug, region = realm, "eu"
-
-        await self._ensure_realm_cache()
-        cache_key = f"{realm_slug}-{region}"
-        if self._realm_cache and cache_key not in self._realm_cache:
-            raise NerpyNotFoundError(get_string(lang, "wow.realm_not_found", realm=realm_slug, region=region.upper()))
-
-        return realm_slug, region
-
-    def _get_retailclient(self, region: str, language: str):
-        if region not in self.regions:
-            raise ValueError(f"Invalid region: {region}. Valid regions are: {', '.join(self.regions)}")
-
-        if language == "de":
-            api_language = WowApiLanguage.DE.value
-        elif region == "eu":
-            api_language = WowApiLanguage.EN_GB.value
-        else:
-            api_language = WowApiLanguage.EN.value
-
-        try:
-            # noinspection PyTypeChecker
-            return RetailClient(
-                client_id=self.client_id,
-                client_secret=self.client_secret,
-                region=Region(region),
-                language=Language(api_language),
-            )
-        except ValueError as ex:
-            raise NerpyInfraException("Failed to initialise WoW API client.") from ex
-
-    def _get_character(self, realm: str, region: str, name: str, language: str) -> tuple[dict, LiteralString]:
-        """Get character profile and media from the WoW API."""
-        api = self._get_retailclient(region, language)
-
-        character = api.character_profile_summary(realmSlug=realm, characterName=name)
-        media = api.character_media(realmSlug=realm, characterName=name)
-        assets = media.get("assets", []) if isinstance(media, dict) else []
-        profile_picture = "".join(asset.get("value") for asset in assets if asset.get("key") == "avatar")
-
-        return character, profile_picture
-
-    # ── Armory command ──────────────────────────────────────────────────
-
-    @app_commands.command(name="armory")
-    async def _wow_armory(
-        self,
-        interaction: Interaction,
-        name: str,
-        realm: str,
-    ):
-        """
-        search for character
-
-        name and realm are required parameters.
-        realm accepts autocomplete suggestions (e.g. "blackrock-eu") or a plain slug (defaults to EU).
-        """
-        try:
-            await interaction.response.defer()
-            lang = self._lang(interaction.guild_id)
-            realm_slug, region = await self._parse_realm(realm, lang)
-            name = name.lower()
-            profile = f"{region}/{realm_slug}/{name}"
-
-            # noinspection PyTypeChecker
-            character, profile_picture = self._get_character(realm_slug, region, name, lang)
-
-            if not isinstance(character, dict) or character.get("code") == 404:
-                raise NerpyNotFoundError(get_string(lang, "wow.armory.not_found"))
-
-            best_keys = get_best_mythic_keys(region, realm_slug, name)
-            rio_score = get_raiderio_score(region, realm_slug, name)
-
-            armory = get_profile_link("armory", profile)
-            raiderio = get_profile_link("raiderio", profile)
-            warcraftlogs = get_profile_link("warcraftlogs", profile)
-            wowprogress = get_profile_link("wowprogress", profile)
-
-            emb = Embed(
-                title=f"{character['name']} | {realm_slug.capitalize()} | {region.upper()} | {character['active_spec']['name']} {character['character_class']['name']} | {character['equipped_item_level']} ilvl",
-                url=armory,
-                color=COLOR_ITEM_LINK,
-                description=f"{character['gender']['name']} {character['race']['name']}",
-            )
-            emb.set_thumbnail(url=profile_picture)
-            emb.add_field(name=get_string(lang, "wow.armory.level"), value=character["level"], inline=True)
-            emb.add_field(name=get_string(lang, "wow.armory.faction"), value=character["faction"]["name"], inline=True)
-            if "guild" in character:
-                emb.add_field(name=get_string(lang, "wow.armory.guild"), value=character["guild"]["name"], inline=True)
-            emb.add_field(name="\u200b", value="\u200b", inline=False)
-
-            if best_keys:
-                keys = ""
-                for key in best_keys:
-                    keys += f"+{key['level']} - {key['dungeon']} - {key['clear_time']}\n"
-
-                emb.add_field(name=get_string(lang, "wow.armory.best_keys"), value=keys, inline=True)
-            if rio_score is not None:
-                emb.add_field(name=get_string(lang, "wow.armory.mplus_score"), value=rio_score, inline=True)
-
-            emb.add_field(name="\u200b", value="\u200b", inline=False)
-            emb.add_field(
-                name=get_string(lang, "wow.armory.external_sites"),
-                value=f"[Raider.io]({raiderio}) | [Armory]({armory}) | [WarcraftLogs]({warcraftlogs}) | [WoWProgress]({wowprogress})",
-                inline=True,
-            )
-
-            await interaction.followup.send(embed=emb)
-        except NerpyUserException as ex:
-            await interaction.followup.send(str(ex), ephemeral=True)
-
-    @_wow_armory.autocomplete("realm")
-    async def _realm_autocomplete_handler(self, interaction: discord.Interaction, current: str):
-        return await self._realm_autocomplete(interaction, current)
 
     # ── Guild News commands ─────────────────────────────────────────────
 
@@ -502,7 +132,7 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
             await interaction.followup.send(str(ex), ephemeral=True)
 
     @_guildnews_setup.autocomplete("realm")
-    async def _guildnews_setup_realm_autocomplete(self, interaction: discord.Interaction, current: str):
+    async def _guildnews_setup_realm_autocomplete(self, interaction, current: str):
         return await self._realm_autocomplete(interaction, current)
 
     async def _config_autocomplete(self, interaction: Interaction, current: str) -> list[app_commands.Choice[int]]:
@@ -694,44 +324,6 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
             await notify_error(self.bot, "Guild news background loop", ex)
         self.bot.log.debug("Stop Guild News Loop!")
 
-    @tasks.loop(hours=1)
-    async def _crafting_cleanup_loop(self):
-        """Delete anchored order messages whose MessageDeleteAt deadline has passed."""
-        self.bot.log.debug("Start Crafting Cleanup Loop!")
-        try:
-            with self.bot.session_scope() as session:
-                pending = CraftingOrder.get_pending_cleanup(session)
-                # Snapshot the fields we need before closing the session
-                to_process = [(o.Id, o.ChannelId, o.OrderMessageId, o.ThreadId) for o in pending]
-
-            for order_id, channel_id, message_id, thread_id in to_process:
-                channel = self.bot.get_channel(channel_id)
-                if channel is None:
-                    self.bot.log.debug("Crafting cleanup: channel %d not found for order #%d", channel_id, order_id)
-                    continue
-                try:
-                    msg = await channel.fetch_message(message_id)
-                    if msg.thread:
-                        try:
-                            await msg.thread.delete()
-                        except discord.HTTPException as ex:
-                            self.bot.log.debug("Crafting cleanup: thread delete failed for order #%d: %s", order_id, ex)
-                    await msg.delete()
-                except discord.NotFound:
-                    pass  # already gone
-                except discord.HTTPException as ex:
-                    self.bot.log.warning("Crafting cleanup: failed to delete message for order #%d: %s", order_id, ex)
-                    continue
-
-                with self.bot.session_scope() as session:
-                    order = CraftingOrder.get_by_id(order_id, session)
-                    if order:
-                        order.MessageDeleteAt = None
-        except Exception as ex:
-            self.bot.log.error("Crafting cleanup loop error: %s", ex)
-            await notify_error(self.bot, "Crafting cleanup background loop", ex)
-        self.bot.log.debug("Stop Crafting Cleanup Loop!")
-
     async def _poll_single_config(self, config_id: int, *, ignore_baseline: bool = False):
         """Poll a single guild news config for activity and mounts."""
         self.bot.log.debug(f"Guild news #{config_id}: starting poll")
@@ -771,10 +363,10 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
 
         api = self._get_retailclient(region, language)
 
-        # ── Phase 1: Guild Activity Feed ────────────────────────────
+        # Phase 1: Guild Activity Feed
         await self._poll_activity(api, cfg_id, wow_guild, realm, last_activity_ts, channel, language)
 
-        # ── Phase 2: Mount Tracking ─────────────────────────────────
+        # Phase 2: Mount Tracking
         if self._track_mounts:
             await self._poll_mounts(api, cfg_id, wow_guild, realm, min_level, active_days, channel, language)
         else:
@@ -1017,7 +609,7 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
 
         if initial_sync:
             self.bot.log.debug(
-                f"Guild news #{config_id}: initial sync — {len(unbaselined)}/{len(candidate_keys)} "
+                f"Guild news #{config_id}: initial sync - {len(unbaselined)}/{len(candidate_keys)} "
                 f"characters not yet baselined, will process all batches"
             )
 
@@ -1122,7 +714,7 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
                         stored = old_entry
 
                 if stored is None:
-                    self.bot.log.debug(f"Guild news #{config_id}: baseline for {char_name} — {len(current_ids)} mounts")
+                    self.bot.log.debug(f"Guild news #{config_id}: baseline for {char_name} - {len(current_ids)} mounts")
                     mount_json = {"ids": sorted(current_ids), "last_count": len(current_ids)}
                     if achievement_points is not None:
                         mount_json["achievement_points"] = achievement_points
@@ -1142,13 +734,13 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
                     removed_ids = known_ids - current_ids
 
                     self.bot.log.debug(
-                        f"Guild news #{config_id}: {char_name} mount diff — "
+                        f"Guild news #{config_id}: {char_name} mount diff - "
                         f"known={len(known_ids)} current={len(current_ids)} new={len(new_ids)} "
                         f"removed={len(removed_ids)} last_count={last_count}"
                     )
 
                     # Churn detection: if IDs both appeared and disappeared with no
-                    # net count increase, it's faction variant ID swapping — not real
+                    # net count increase, it's faction variant ID swapping - not real
                     # new mounts.
                     if removed_ids and new_ids:
                         net_new = len(current_ids) - last_count
@@ -1163,7 +755,7 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
                     if not should_update_mount_set(last_count, len(current_ids)):
                         self.bot.log.warning(
                             f"Guild news #{config_id}: {char_name} mount count dropped "
-                            f"(last_count={last_count} current={len(current_ids)}) — "
+                            f"(last_count={last_count} current={len(current_ids)}) - "
                             f"likely degraded API response, skipping update"
                         )
                         total_stats["skipped_degraded"] += 1
@@ -1238,7 +830,7 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
                     stored.LastChecked = datetime.now(UTC)
                     cycle_new_mounts[(char_name, char_realm)] = new_ids
 
-        # Process batches — loop through all during initial sync, single batch otherwise
+        # Process batches - loop through all during initial sync, single batch otherwise
         with self.bot.session_scope() as session:
             config = session.query(WowGuildNewsConfig).filter(WowGuildNewsConfig.Id == config_id).first()
             if not config:
@@ -1272,14 +864,14 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
             await asyncio.gather(*[_check_character(c) for c in batch])
 
             self.bot.log.debug(
-                f"Guild news #{config_id}: batch #{batch_num} done — "
+                f"Guild news #{config_id}: batch #{batch_num} done - "
                 f"checked={total_stats['checked']}, baselined={total_stats['baselined']}, "
                 f"new_mounts={total_stats['new_mounts']}, "
                 f"skipped_inactive={total_stats['skipped_inactive']}, skipped_error={total_stats['skipped_error']}, "
                 f"skipped_degraded={total_stats['skipped_degraded']}, skipped_404={total_stats['skipped_404']}"
             )
 
-            # Rate limited — stop immediately, resume from current offset next cycle
+            # Rate limited - stop immediately, resume from current offset next cycle
             if rate_limited.is_set():
                 self.bot.log.warning(
                     f"Guild news #{config_id}: stopping mount poll due to rate limit, "
@@ -1292,7 +884,7 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
             if not initial_sync:
                 break
 
-            # Full rotation completed — stop if no new baselines were added this rotation
+            # Full rotation completed - stop if no new baselines were added this rotation
             if offset == start_offset or (offset == 0 and start_offset >= len(candidate_list)):
                 if total_stats["baselined"] == baselined_before:
                     self.bot.log.debug(
@@ -1300,7 +892,7 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
                         f"remaining characters are inactive/inaccessible"
                     )
                     break
-                # New rotation — reset the counter
+                # New rotation - reset the counter
                 baselined_before = total_stats["baselined"]
                 start_offset = offset
 
@@ -1334,429 +926,8 @@ class WorldofWarcraft(NerpyBotCog, GroupCog, group_name="wow"):
 
         if initial_sync and not rate_limited.is_set():
             self.bot.log.debug(
-                f"Guild news #{config_id}: initial sync finished in {batch_num} batches — "
+                f"Guild news #{config_id}: initial sync finished in {batch_num} batches - "
                 f"baselined={total_stats['baselined']}, skipped_inactive={total_stats['skipped_inactive']}, "
                 f"skipped_error={total_stats['skipped_error']}, skipped_degraded={total_stats['skipped_degraded']}, "
                 f"skipped_404={total_stats['skipped_404']}"
             )
-
-    # ── Crafting Order commands ────────────────────────────────────────
-
-    @craftingorder.command(name="create")
-    @checks.has_permissions(manage_channels=True)
-    @app_commands.describe(
-        channel="Channel where the board embed will be posted",
-        description="Description shown on the board embed (opens a modal if omitted)",
-        roles="Profession role mentions separated by spaces or commas",
-        description_message="Message ID or link whose text becomes the description (message is deleted)",
-    )
-    @app_commands.rename(description_message="description-message")
-    async def _craftingorder_create(
-        self,
-        interaction: Interaction,
-        channel: TextChannel,
-        roles: str,
-        description: str | None = None,
-        description_message: str | None = None,
-    ):
-        """create a crafting order board in a channel [manage_channels]"""
-        lang = self._lang(interaction.guild_id)
-
-        # Resolve description from message reference if provided
-        if description_message:
-            from utils.helpers import fetch_message_content
-
-            content, error = await fetch_message_content(
-                self.bot,
-                description_message,
-                channel,
-                interaction,
-                lang,
-                key_prefix="wow.craftingorder.fetch_description",
-            )
-            if error:
-                await interaction.response.send_message(error, ephemeral=True)
-                return
-            description = content
-
-        if not description:
-            # No description provided — show a modal to collect it
-            modal = _BoardDescriptionModal(self.bot, lang, mode="create", channel=channel, roles=roles)
-            await interaction.response.send_modal(modal)
-            return
-
-        if description and not description_message:
-            # Inline description provided — show modal pre-filled for review
-            modal = _BoardDescriptionModal(
-                self.bot, lang, mode="create", channel=channel, roles=roles, default_text=description
-            )
-            await interaction.response.send_modal(modal)
-            return
-
-        await interaction.response.defer(ephemeral=True)
-        await self.finish_board_create(interaction, channel, roles, description, lang)
-
-    @staticmethod
-    def _parse_role_ids(roles_str: str, guild: discord.Guild) -> list[int]:
-        """Parse role mentions/IDs from a space/comma-separated string."""
-        role_ids = []
-        for part in roles_str.replace(",", " ").split():
-            part = part.strip("<@&>")
-            if part.isdigit():
-                role = guild.get_role(int(part))
-                if role:
-                    role_ids.append(role.id)
-        return role_ids
-
-    @staticmethod
-    def _auto_match_roles(guild: discord.Guild, role_ids: list[int], session) -> tuple[list[int], list[int]]:
-        """Auto-match Discord roles to Blizzard profession IDs.
-
-        Returns (mapped_role_ids, unmapped_role_ids).
-        Creates CraftingRoleMapping rows for each match.
-        """
-        mapped = []
-        unmapped = []
-        for role_id in role_ids:
-            role = guild.get_role(role_id)
-            if not role:
-                unmapped.append(role_id)
-                continue
-
-            matched = False
-            for prof_name, prof_id in CRAFTING_PROFESSIONS.items():
-                if prof_name.lower() in role.name.lower():
-                    session.add(CraftingRoleMapping(GuildId=guild.id, RoleId=role_id, ProfessionId=prof_id))
-                    mapped.append(role_id)
-                    matched = True
-                    break
-
-            if not matched:
-                unmapped.append(role_id)
-
-        return mapped, unmapped
-
-    async def finish_board_create(
-        self, interaction: Interaction, channel: TextChannel, roles: str, description: str, lang: str
-    ):
-        """Shared board creation logic used by both the command and the description modal."""
-        role_ids = self._parse_role_ids(roles, interaction.guild)
-
-        if not role_ids:
-            await interaction.followup.send(get_string(lang, "wow.craftingorder.create.no_roles"), ephemeral=True)
-            return
-
-        validate_channel_permissions(channel, interaction.guild, "send_messages", "embed_links", "manage_threads")
-
-        with self.bot.session_scope() as session:
-            existing = CraftingBoardConfig.get_by_guild(interaction.guild_id, session)
-            if existing:
-                existing_channel = interaction.guild.get_channel(existing.ChannelId)
-                ch_mention = existing_channel.mention if existing_channel else f"#{existing.ChannelId}"
-                await interaction.followup.send(
-                    get_string(lang, "wow.craftingorder.create.already_exists", channel=ch_mention),
-                    ephemeral=True,
-                )
-                return
-
-            config = CraftingBoardConfig(
-                GuildId=interaction.guild_id,
-                ChannelId=channel.id,
-                Description=description,
-                BoardVersion=CURRENT_BOARD_VERSION,
-            )
-            session.add(config)
-            session.flush()
-
-            # Auto-match roles to Blizzard professions
-            mapped, unmapped = self._auto_match_roles(interaction.guild, role_ids, session)
-
-        # Build embed and view then send to Discord outside the session so the DB
-        # connection is not held open during the HTTP call.
-        from modules.views.crafting_order import CraftingBoardView
-
-        embed = discord.Embed(
-            title=get_string(lang, "wow.craftingorder.board_title"),
-            description=description,
-            color=discord.Color.gold(),
-        )
-        embed.set_footer(text=get_string(lang, "wow.craftingorder.board_footer"))
-
-        view = CraftingBoardView(
-            self.bot,
-            label=get_string(lang, "wow.craftingorder.create_button"),
-            housing_label=get_string(lang, "wow.craftingorder.housing_button"),
-        )
-        try:
-            msg = await channel.send(embed=embed, view=view)
-        except discord.HTTPException as exc:
-            # Board config was committed above; clean it up so the guild can retry.
-            with self.bot.session_scope() as session:
-                orphaned = CraftingBoardConfig.get_by_guild(interaction.guild_id, session)
-                if orphaned is not None:
-                    session.delete(orphaned)
-                CraftingRoleMapping.delete_by_guild(interaction.guild_id, session)
-            self.bot.log.error("Failed to post crafting board embed (guild=%d): %s", interaction.guild_id, exc)
-            await interaction.followup.send(
-                get_string(lang, "wow.craftingorder.create.send_failed", channel=channel.mention), ephemeral=True
-            )
-            return
-
-        with self.bot.session_scope() as session:
-            config = CraftingBoardConfig.get_by_guild(interaction.guild_id, session)
-            if config is not None:
-                config.BoardMessageId = msg.id
-
-        await interaction.followup.send(
-            get_string(lang, "wow.craftingorder.create.success", channel=channel.mention), ephemeral=True
-        )
-
-        if unmapped:
-            await self._send_manual_mapping_view(interaction, unmapped, lang)
-
-    @craftingorder.command(name="remove")
-    @checks.has_permissions(manage_channels=True)
-    async def _craftingorder_remove(self, interaction: Interaction):
-        """remove the crafting order board [manage_channels]"""
-        await interaction.response.defer(ephemeral=True)
-        lang = self._lang(interaction.guild_id)
-
-        with self.bot.session_scope() as session:
-            config = CraftingBoardConfig.delete_by_guild(interaction.guild_id, session)
-            if config is None:
-                await interaction.followup.send(get_string(lang, "wow.craftingorder.remove.not_found"), ephemeral=True)
-                return
-            CraftingRoleMapping.delete_by_guild(interaction.guild_id, session)
-            channel_id = config.ChannelId
-            message_id = config.BoardMessageId
-
-        # Try to delete the board embed
-        try:
-            channel = interaction.guild.get_channel(channel_id)
-            if channel and message_id:
-                msg = await channel.fetch_message(message_id)
-                if msg.thread:
-                    await msg.thread.delete()
-                await msg.delete()
-        except discord.HTTPException:
-            self.bot.log.debug("Failed to delete crafting board message (channel=%s, msg=%s)", channel_id, message_id)
-
-        await interaction.followup.send(get_string(lang, "wow.craftingorder.remove.success"), ephemeral=True)
-
-    @craftingorder.command(name="edit")
-    @checks.has_permissions(manage_channels=True)
-    @app_commands.describe(roles="New role mentions (optional — re-runs auto-matching)")
-    async def _craftingorder_edit(self, interaction: Interaction, roles: str | None = None):
-        """edit the crafting order board [manage_channels]"""
-        lang = self._lang(interaction.guild_id)
-
-        with self.bot.session_scope() as session:
-            config = CraftingBoardConfig.get_by_guild(interaction.guild_id, session)
-            if config is None:
-                await interaction.response.send_message(
-                    get_string(lang, "wow.craftingorder.edit.not_found"), ephemeral=True
-                )
-                return
-            current_description = config.Description
-
-        # If new roles provided, update the mapping before showing the modal
-        unmapped: list[int] = []
-        if roles:
-            role_ids = self._parse_role_ids(roles, interaction.guild)
-            if role_ids:
-                with self.bot.session_scope() as session:
-                    CraftingRoleMapping.delete_by_guild(interaction.guild_id, session)
-                    _mapped, unmapped = self._auto_match_roles(interaction.guild, role_ids, session)
-
-        # Show modal pre-filled with current description
-        modal = _BoardDescriptionModal(self.bot, lang, mode="edit", default_text=current_description)
-        modal._unmapped = unmapped
-        await interaction.response.send_modal(modal)
-
-    async def finish_board_edit(
-        self, interaction: Interaction, new_description: str, lang: str, unmapped: list[int] | None = None
-    ):
-        """Update the board description and edit the embed in-place."""
-        with self.bot.session_scope() as session:
-            config = CraftingBoardConfig.get_by_guild(interaction.guild_id, session)
-            if config is None:
-                await interaction.followup.send(get_string(lang, "wow.craftingorder.edit.not_found"), ephemeral=True)
-                return
-            config.Description = new_description
-            channel_id = config.ChannelId
-            message_id = config.BoardMessageId
-
-        # Edit the board embed in-place and refresh the view with the housing button.
-        # Pre-set True when there is no message to update — description was saved, nothing to fail.
-        embed_updated = not message_id
-        if message_id:
-            try:
-                from modules.views.crafting_order import CraftingBoardView
-
-                channel = interaction.guild.get_channel(channel_id) or await interaction.guild.fetch_channel(channel_id)
-                msg = await channel.fetch_message(message_id)
-                embed = msg.embeds[0] if msg.embeds else discord.Embed(color=discord.Color.gold())
-                embed.description = new_description
-                new_view = CraftingBoardView(
-                    self.bot,
-                    label=get_string(lang, "wow.craftingorder.create_button"),
-                    housing_label=get_string(lang, "wow.craftingorder.housing_button"),
-                )
-                await msg.edit(embed=embed, view=new_view)
-                embed_updated = True
-            except discord.HTTPException as exc:
-                self.bot.log.warning("Failed to edit board embed (channel=%s, msg=%s): %s", channel_id, message_id, exc)
-
-        if embed_updated:
-            await interaction.followup.send(get_string(lang, "wow.craftingorder.edit.success"), ephemeral=True)
-        else:
-            await interaction.followup.send(get_string(lang, "wow.craftingorder.edit.embed_failed"), ephemeral=True)
-
-        if unmapped:
-            await self._send_manual_mapping_view(interaction, unmapped, lang)
-
-    @Cog.listener("on_guild_language_changed")
-    async def _on_guild_language_changed(self, guild_id: int, new_lang: str) -> None:
-        """Refresh persistent WoW embeds when a guild's language preference changes."""
-        await self._refresh_crafting_board(guild_id, new_lang)
-        await self._refresh_active_orders(guild_id, new_lang)
-
-    async def _refresh_crafting_board(self, guild_id: int, new_lang: str) -> None:
-        """Edit the crafting board embed and view in-place with the new language."""
-        from modules.views.crafting_order import CraftingBoardView
-
-        with self.bot.session_scope() as session:
-            config = CraftingBoardConfig.get_by_guild(guild_id, session)
-            if config is None or not config.BoardMessageId or not config.ChannelId:
-                return
-            channel_id = config.ChannelId
-            message_id = config.BoardMessageId
-            description = config.Description
-
-        try:
-            guild = self.bot.get_guild(guild_id)
-            channel = guild.get_channel(channel_id) if guild else None
-            if channel is None:
-                channel = await self.bot.fetch_channel(channel_id)
-            msg = await channel.fetch_message(message_id)
-            embed = discord.Embed(
-                title=get_string(new_lang, "wow.craftingorder.board_title"),
-                description=description,
-                color=discord.Color.gold(),
-            )
-            embed.set_footer(text=get_string(new_lang, "wow.craftingorder.board_footer"))
-            view = CraftingBoardView(
-                self.bot,
-                label=get_string(new_lang, "wow.craftingorder.create_button"),
-                housing_label=get_string(new_lang, "wow.craftingorder.housing_button"),
-            )
-            await msg.edit(embed=embed, view=view)
-            self.bot.log.info("wow: refreshed crafting board embed for guild %d (lang=%s)", guild_id, new_lang)
-        except (discord.NotFound, discord.Forbidden):
-            self.bot.log.warning("wow: crafting board message not accessible for guild %d — skipping refresh", guild_id)
-        except discord.HTTPException as exc:
-            self.bot.log.warning("wow: failed to refresh crafting board for guild %d: %s", guild_id, exc)
-
-    async def _refresh_active_orders(self, guild_id: int, new_lang: str) -> None:
-        """Edit each active crafting order card embed and view with the new language."""
-        from modules.views.crafting_order import build_order_embed, build_order_view
-
-        with self.bot.session_scope() as session:
-            orders = CraftingOrder.get_active_by_guild(guild_id, session)
-            order_data = [
-                (order.Id, order.ChannelId, order.OrderMessageId)
-                for order in orders
-                if order.OrderMessageId and order.ChannelId
-            ]
-
-        guild = self.bot.get_guild(guild_id)
-        for i, (order_id, channel_id, message_id) in enumerate(order_data):
-            try:
-                channel = guild.get_channel(channel_id) if guild else None
-                if channel is None:
-                    channel = await self.bot.fetch_channel(channel_id)
-                msg = await channel.fetch_message(message_id)
-                with self.bot.session_scope() as session:
-                    order = CraftingOrder.get_by_id(order_id, session)
-                    if order is None:
-                        continue
-                    embed = build_order_embed(order, guild, new_lang)
-                    view = build_order_view(order.Id, order.Status, new_lang)
-                await msg.edit(embed=embed, view=view)
-                self.bot.log.info("wow: refreshed order #%d embed for guild %d (lang=%s)", order_id, guild_id, new_lang)
-            except (discord.NotFound, discord.Forbidden):
-                self.bot.log.warning(
-                    "wow: order #%d message not accessible for guild %d — skipping refresh", order_id, guild_id
-                )
-            except discord.HTTPException as exc:
-                self.bot.log.warning("wow: failed to refresh order #%d for guild %d: %s", order_id, guild_id, exc)
-            if i < len(order_data) - 1:
-                await asyncio.sleep(1)
-
-    async def _send_manual_mapping_view(self, interaction: Interaction, unmapped: list[int], lang: str):
-        """Send an ephemeral followup with Select dropdowns for manual profession mapping."""
-        from modules.views.crafting_order import ManualProfessionMappingView
-
-        unmapped_roles = [
-            (rid, (interaction.guild.get_role(rid).name if interaction.guild.get_role(rid) else str(rid)))
-            for rid in unmapped
-        ]
-        content = get_string(lang, "wow.craftingorder.manual_map.description")
-        view = ManualProfessionMappingView(self.bot, interaction.guild_id, unmapped_roles, lang)
-        await interaction.followup.send(content, view=view, ephemeral=True)
-
-
-class _BoardDescriptionModal(discord.ui.Modal):
-    """Modal for collecting/editing board description.
-
-    Supports two modes:
-    - "create": calls finish_board_create after submission
-    - "edit": updates existing board config and edits the embed in-place
-    """
-
-    description_input = discord.ui.TextInput(
-        label="Board Description",
-        style=discord.TextStyle.paragraph,
-        max_length=4000,
-        required=True,
-    )
-
-    def __init__(
-        self,
-        bot,
-        lang: str,
-        *,
-        mode: str = "create",
-        channel: TextChannel = None,
-        roles: str = None,
-        default_text: str = None,
-    ):
-        super().__init__(title=get_string(lang, "wow.craftingorder.create.modal_title"))
-        self.bot = bot
-        self.lang = lang
-        self.mode = mode
-        self.channel = channel
-        self.roles = roles
-        self.description_input.placeholder = get_string(lang, "wow.craftingorder.create.modal_description")
-        if default_text:
-            self.description_input.default = default_text
-
-    async def on_submit(self, interaction: Interaction):
-        await interaction.response.defer(ephemeral=True)
-        cog = self.bot.get_cog("WorldofWarcraft")
-
-        if self.mode == "create":
-            await cog.finish_board_create(
-                interaction, self.channel, self.roles, self.description_input.value.strip(), self.lang
-            )
-        else:
-            unmapped = getattr(self, "_unmapped", [])
-            await cog.finish_board_edit(interaction, self.description_input.value.strip(), self.lang, unmapped=unmapped)
-
-
-async def setup(bot):
-    """adds this module to the bot"""
-    if "wow" in bot.config:
-        await bot.add_cog(WorldofWarcraft(bot))
-    else:
-        raise NerpyInfraException("Config not found.")
